Organize code in specified directories
